+++ src/App.tsx
@@ -2,7 +2,7 @@ import "./App.css";
 import { useState } from "react";
 import { AppStateContext } from "././context/AppStateContext";
 import { Window as KeplrWindow } from "@keplr-wallet/types";
-import MainPage from "./components/MainPage/MainPage";
+import CreatePage from "./components/CreatePage/CreatePage";
 import Footer from "./components/Footer/Footer";
 import LegalPage from "./components/LegalPage/LegalPage";
 import { HashRouter as Router, Routes, Route } from "react-router-dom";
@@ -10,9 +10,8 @@ import { ToastContainer } from "react-toastify";
 import "react-toastify/dist/ReactToastify.css";
 import Header from "./components/Header/Header";
 import WelcomePage from "./components/WelcomePage/WelcomePage";
-import NewButton from "./components/newButton/newButton";
 import Swap from "./components/Swap/Swap";
-import ManageAssets from "./components/ManageAssets/ManageAssets";
+import MyWallet from "./components/MyWalletPage/MyWallet";
 import License from "./components/LicensePage/LicensePage";
 import { GrazProvider } from "graz";
 import { CustomChains } from "../src/utils/config";
@@ -37,11 +36,10 @@ function App() {
               <Routes>
                 <Route path="/" element={<WelcomePage />}></Route>
                 <Route path="/legalinfo" element={<LegalPage />}></Route>
-                <Route path="/License" element={<License />}></Route>
-                <Route path="/create" element={<MainPage />}></Route>
-                <Route path="/newBT" element={<NewButton />}></Route>
+                <Route path="/license" element={<License />}></Route>
+                <Route path="/create" element={<CreatePage />}></Route>
                 <Route path="/swap" element={<Swap />}></Route>
-                <Route path="/manage-assets" element={<ManageAssets />}></Route>
+                <Route path="/my-wallet" element={<MyWallet />}></Route>
               </Routes>
               <Footer />
             </Router>
