+++ src/components/CreatePage/Form/Form.module.css
@@ -5,8 +5,8 @@
   position: relative;
   width: 100%;
   height: 100%;
-  background: rgba(124, 196, 219, 0.05);
-  border: solid 1px rgba(115, 237, 201, 0.2);
+  background: #12262d;
+  border: solid 1px rgba(115, 237, 201, 0.5);
   border-radius: 0px 16px 16px;
   backdrop-filter: blur(10px);
 }
@@ -29,7 +29,7 @@
   display: flex;
   align-items: center;
   justify-content: center;
-  gap: 30px;
+  gap: 15px;
 }
 
 .inputComponent {
@@ -60,6 +60,16 @@
 }
 
 .connectButton {
-  font-size: 16px;
-  font-weight: 500;
+  font-family: "Bai Jamjuree", sans-serif;
+  font-weight: 700;
+  font-size: 32px;
+}
+
+.article {
+  margin: 5px 0px 10px 0px;
+  font-style: normal;
+  font-weight: 400;
+  font-size: 14px;
+  line-height: 18px;
+  color: rgba(246, 248, 254, 0.5);
 }
