+++ src/components/CreatePage/NTT/NTT.module.css
@@ -1,10 +1,10 @@
-.nft {
+.ntt {
   display: flex;
   flex-direction: column;
   box-sizing: border-box;
   width: 100%;
   height: 100%;
-  background: rgba(124, 196, 219, 0.05);
+  background: #12262d;
   border: solid 1px rgba(115, 237, 201, 0.2);
   border-radius: 0px 16px 16px;
   backdrop-filter: blur(10px);
