+++ src/components/CreatePage/Tabs/Tabs.module.css
@@ -9,10 +9,10 @@
 .tabLabel {
   min-width: 144px;
   padding: 7px 0px;
-  background-color: #1a2d36;
+  background-color: #12262d;
   border-style: solid solid none;
   border-width: 1px;
-  border-color: rgba(115, 237, 201, 0.2);
+  border-color: rgba(115, 237, 201, 0.5);
   border-radius: 12px 12px 0px 0px;
   color: #E2FB5F;
   font-weight: 600;
@@ -28,7 +28,8 @@
 
 .tabLabel__selected {
   padding: 9px 0px;
-  background-color: rgb(16, 33, 40, 0.7);
+  background-color: rgba(20, 35, 42, 0.85);
+  border-color: rgba(115, 237, 201, 0.4);
   border-style: solid;
   border-width: 1px;
   font-weight: 400;
@@ -38,7 +39,7 @@
 }
 
 .tabLabel__selected:hover {
-  background-color: rgb(16, 33, 40, 0.7);
+  background-color: rgb(25, 44, 52);
   font-weight: 500;
   color: #E2FB5F;
   backdrop-filter: blur(5px);
