+++ src/components/Footer/Footer.module.css
@@ -9,7 +9,7 @@
 }
 
 .link {
-  color: #535e6b;
+  color: rgba(246, 248, 254, 0.6);
   width: fit-content;
   font-family: 'Bai Jamjuree';
   font-style: normal;
