+++ src/components/LegalPage/LegalPage.module.css
@@ -63,7 +63,7 @@
 .lastUpd {
   font-style: normal;
   font-size: 14px;
-  color: rgba(246, 248, 254, 0.2);
+  color: rgba(115, 237, 201, 0.3);
   padding-top: 30px;
   padding-bottom: 15px;
 }
