+++ src/components/ManageAssets/ManageAssets.module.css
@@ -1,126 +1,116 @@
-.nav {
+.manageTok {
   display: flex;
   flex-direction: column;
-  align-items: center;
-  margin: 10px auto auto auto;
+  background: rgb(25, 44, 52, 0.50);
+  backdrop-filter: blur(10px);
+  border-radius: 16px;
+  min-width: 262px;
+  min-height: 262px;
+  height: 100%;
+  margin-left: 9px;
+  z-index: 100;
 }
+
+.highlighter {
+  width: 100%;
+  border-radius: 16px 16px 0px 0px;
+  background: #e2fb5f;
+}
+
 .name {
-  position: absolute;
-  color: rgba(67, 178, 91, 0.15);
-  font-size: 167px;
-  font-weight: bold;
-  text-align: center;
-  margin-top: 50px;
-}
-.border {
-  z-index: 5;
-  display: block;
-  margin-top: 170px;
-  border-radius: 10px !important;
-  overflow: hidden !important;
-}
-.test {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-.tableMA {
-  border-radius: 10px;
-  border-collapse: collapse;
-}
-
-.thead {
-  background-color: rgba(22, 40, 51, 0.9);
-}
-.thead th {
-  color: #43b25b;
-  font-size: 14px;
-  font-weight: 200;
-  padding-top: 9px;
-  padding-bottom: 11px;
-}
-.tokens {
-  padding-left: 76px;
-  padding-right: 62px;
-}
-.balance {
-  padding-right: 188px;
-}
-.other {
-  padding-right: 17px;
-  padding-left: 17px;
-}
-.nameToken {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-.boot {
-  color: white;
-  font-size: 20px;
-}
-.bostorm {
-  color: #43b25b;
-  font-size: 12px;
-  padding-right: 5px;
-}
-
-.circle {
-  width: 32px;
-  margin-right: 13px;
-}
-.price {
-  color: white;
+  color: #060F19;
+  font-family: 'Bai Jamjuree';
+  font-style: normal;
+  font-weight: 500;
   font-size: 24px;
-  text-align: start;
+  line-height: 30px;
+  text-transform: lowercase;
+  padding: 12px 24px;
 }
-.menu {
-  background-color: rgba(22, 40, 51, 0.5);
-  backdrop-filter: blur(15px);
+
+.tokens {
+  padding: 12px 24px 26px 24px;
 }
-.exchangeField {
+
+.tokenList {
+  width: 100%;
+  margin-bottom: 15px;
+}
+
+.info {
+  font-weight: 400;
+  font-size: 12px;
+  color: rgba(246, 248, 254, 0.6);
+  margin: 8px 0px 5px 0px;
+}
+
+.info2 {
+  padding: 75px 0px 90px;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  font-weight: 400;
+  font-size: 14px;
+  line-height: 18px;
+  text-align: center;
+  color: rgba(115, 237, 201, 0.3);
+}
+
+.inputs {
   display: flex;
   flex-direction: column;
-  color: #e2fb5f;
-  font-size: 12px;
-  font-weight: 200;
+  align-items: flex-start;
 }
 
-.buttonY {
-  margin-left: 50px;
-}
-.exchange {
-  display: flex;
-  background-color: rgba(22, 40, 51, 0.5);
-  padding: 10px 0px 10px 15px;
-}
-.field {
-  border: 1px solid rgba(67, 178, 91, 0.35);
-  border-radius: 10px;
-  background-color: rgba(22, 40, 51, 0.1);
-  padding: 13px 33px 13px 10px;
-  outline: none;
-  margin: 5px 14px 5px 0px;
-  color: #e2fb5f;
+.addContract {
+  box-sizing: border-box;
+  height: 42px;
+  width: 100%;
   font-family: "Bai Jamjuree", sans-serif;
+  font-weight: 400;
+  font-size: 16px;
+  color: #F6F8FE;
+  background: transparent;
+  border: 1px solid rgba(115, 237, 201, 0.5);
+  border-radius: 8px;
+  padding-left: 10px;
 }
 
-.add {
+.addTokenButton {
+  margin-top: 16px;
+}
+
+.error {
   display: flex;
-  justify-content: center;
-  background-color: rgba(22, 40, 51, 0.5);
-  padding-top: 21px;
-  padding-bottom: 21px;
+  align-content: center;
+  color: rgba(246, 248, 254, 0.6);
+  font-family: 'Bai Jamjuree';
+  font-style: normal;
+  font-weight: 400;
+  font-size: 14px;
+  text-align: center;
+  padding-top: 7px;
 }
-.tokenSend {
-  margin-top: 10px;
-  margin-left: 15px;
-  padding: 10px 55px;
+
+@keyframes rotate {
+  100% {
+    transform: rotate(360deg);
+  }
 }
-.addToken {
-  margin-left: 0px;
-  padding: 10px 45px;
-}
-.lg {
-  font-weight: 100;
+
+@keyframes dash {
+  0% {
+    stroke-dasharray: 1, 150;
+    stroke-dashoffset: 0;
+  }
+
+  50% {
+    stroke-dasharray: 90, 150;
+    stroke-dashoffset: -35;
+  }
+
+  100% {
+    stroke-dasharray: 90, 150;
+    stroke-dashoffset: -124;
+  }
 }
