+++ src/components/ManageAssets/Token/Token.module.css
@@ -4,6 +4,7 @@
   margin-top: 12px;
 }
 
+
 .tokenTitle {
   display: flex;
   flex-direction: row;
@@ -23,6 +24,12 @@
   border: none;
   cursor: pointer;
   font-family: "Bai Jamjuree", sans-serif;
+  padding: 12px 10px;
+  border-radius: 8px;
+}
+
+.cashName:hover {
+  background: rgba(67, 151, 178, 0.1);
 }
 
 .x {
@@ -31,19 +38,20 @@
   color: #43b25b;
   padding: 0px;
   cursor: pointer;
-}
-
-.children {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
+  border-radius: 50%;
 }
 
 .label {
-  font-weight: 700;
-  color: #43b25b;
-  font-size: 10px;
-  margin: 0px 5px;
+  font-weight: 400;
+  font-size: 12px;
+  color: rgba(246, 248, 254, 0.6);
+  margin: 8px 0px 5px 0px;
+}
+
+.children {
+  width: 100%;
+  display: flex;
+  flex-direction: column;
 }
 
 .logo {
@@ -57,30 +65,31 @@
 
 .image {
   margin-left: 5px;
+  width: 10px;
 }
 
 .balance {
+  font-weight: 400;
+  font-size: 14px;
   margin-left: auto;
 }
 
 .addContract {
-  height: 44px;
-  color: #e2fb5f;
-  font-weight: 400;
   box-sizing: border-box;
-  font-size: 16px;
-  line-height: 28px;
-  background: transparent;
-  border-radius: 10px;
-  padding: 8px;
-  border: 1px solid rgba(67, 178, 91, 0.35);
-  outline: none;
-  margin-bottom: 12px;
+  height: 42px;
+  width: 100%;
   font-family: "Bai Jamjuree", sans-serif;
+  font-weight: 400;
+  font-size: 16px;
+  color: #F6F8FE;
+  background: transparent;
+  border: 1px solid rgba(115, 237, 201, 0.5);
+  border-radius: 8px;
+  padding-left: 10px;
 }
 
 .info {
-  color: #43b25b;
+  color: rgba(246, 248, 254, 0.6);
   font-size: 14px;
   padding: 1px 25px 5px 25px;
 }
