+++ src/components/ManageAssets/TokenSender/TokenSender.module.css
@@ -18,6 +18,12 @@
   background-color: transparent;
   border: none;
   font-family: "Bai Jamjuree", sans-serif;
+  padding: 12px 10px;
+  border-radius: 8px;
+}
+
+.cashName:hover {
+  background: rgba(67, 151, 178, 0.1);
 }
 
 .token {
@@ -28,27 +34,28 @@
 
 .image {
   margin-right: auto;
+  width: 10px;
 }
 
 .addContract {
-  color: #e2fb5f;
-  font-weight: 400;
   box-sizing: border-box;
-  font-size: 16px;
-  line-height: 28px;
-  background: transparent;
-  border-radius: 10px;
-  padding: 8px;
-  border: 1px solid rgba(67, 178, 91, 0.35);
-  outline: none;
+  height: 42px;
+  width: 100%;
   font-family: "Bai Jamjuree", sans-serif;
+  font-weight: 400;
+  font-size: 16px;
+  color: #F6F8FE;
+  background: transparent;
+  border: 1px solid rgba(115, 237, 201, 0.5);
+  border-radius: 8px;
+  padding-left: 10px;
 }
 
 .label {
-  font-weight: 700;
-  color: #43b25b;
-  font-size: 10px;
-  margin: 14px 5px 3px 5px;
+  font-weight: 400;
+  font-size: 12px;
+  color: rgba(246, 248, 254, 0.6);
+  margin: 8px 0px 5px 0px;
 }
 
 .children {
@@ -63,13 +70,12 @@
   align-self: center;
   font-style: normal;
   font-weight: 700;
-  color: #43b25b;
+  color: rgba(246, 248, 254, 0.6);
   font-size: 14px;
   margin: 5px 4px 5px 0px;
   padding: 1px 25px 5px 25px;
 }
 
 .tokenButton {
-  margin-top: 20px;
-  padding: 12px 90px;
+  margin-top: 16px;
 }
