+++ src/components/ManageAssets/TokenSender/TokenSender.tsx
@@ -1,10 +1,10 @@
 import React from "react";
 import styles from "./TokenSender.module.css";
-import Button from "../Button/Button";
+import Button from "../../Button/Button";
 import { useState } from "react";
-import { sendBoot } from "../../utils/wallet";
+import { sendBoot } from "../../../utils/wallet";
 import { coins } from "@cosmjs/stargate";
-import collapse_arrow from "../../assets/collapse_arrow.svg";
+import collapse_arrow from "../../../assets/plus.svg";
 import { useBalances, useAccount } from "graz";
 
 const sendBalance = {
@@ -86,7 +86,7 @@ function TokenSender() {
             onChange={handleSendChange("amount")}
           />
           <Button
-            color="white"
+            color="sendButton"
             type="button"
             className={styles.tokenButton}
             onClick={(e: any) => getBootSent(balance.recepient, balance.amount)}
