--- src/components/ManageTokens/ManageTok.module.css
@@ -1,133 +1,0 @@
-.manageTok {
-  display: flex;
-  flex-direction: column;
-  background: rgba(124, 196, 219, 0.05);
-  border: 1px solid rgba(67, 151, 178, 0.1);
-  backdrop-filter: blur(10px);
-  border-radius: 16px;
-  min-width: 262px;
-  min-height: 262px;
-  height: 100%;
-  margin-left: 9px;
-  z-index: 100;
-}
-
-.highlighter {
-  width: 100%;
-  border-radius: 16px 16px 0px 0px;
-  background: rgba(9, 15, 21, 0.35);
-}
-
-.name {
-  color: #e2fb5f;
-  font-family: 'Bai Jamjuree';
-  font-style: normal;
-  font-weight: 500;
-  font-size: 24px;
-  line-height: 30px;
-  text-transform: lowercase;
-  padding: 12px 24px;
-}
-
-.tokens {
-  padding: 12px 24px 26px 24px;
-}
-
-.tokenList {
-  width: 100%;
-  margin-bottom: 15px;
-}
-
-.info {
-  font-weight: 400;
-  font-size: 12px;
-  color: rgba(246, 248, 254, 0.6);
-  margin: 8px 0px 5px 0px;
-}
-
-.info2 {
-  padding: 75px 0px 90px;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  font-weight: 400;
-  font-size: 14px;
-  line-height: 18px;
-  text-align: center;
-  color: rgba(115, 237, 201, 0.3);
-}
-
-.inputs {
-  display: flex;
-  flex-direction: column;
-  align-items: flex-start;
-}
-
-.addContract {
-  box-sizing: border-box;
-  height: 42px;
-  width: 100%;
-  font-family: "Bai Jamjuree", sans-serif;
-  font-weight: 400;
-  font-size: 16px;
-  color: #F6F8FE;
-  background: transparent;
-  border: 1px solid rgba(115, 237, 201, 0.3);
-  border-radius: 8px;
-  padding-left: 10px;
-}
-
-.spinner {
-  display: flex;
-  justify-content: center;
-  animation: rotate 0.1s linear infinite;
-  width: 25px;
-  height: 25px;
-  padding: 0px 8px;
-}
-
-.path {
-  display: flex;
-  stroke: #d3ff54;
-  stroke-linecap: round;
-  animation: dash 0.1s ease-in-out infinite;
-  align-self: center;
-}
-
-.addTokenButton {
-  margin-top: 16px;
-}
-
-.error {
-  display: flex;
-  justify-content: center;
-  color: #e2fb5f;
-  font-family: 'Bai Jamjuree';
-  font-style: normal;
-  font-weight: 500;
-  font-size: 14px;
-  padding-top: 4px;
-}
-
-@keyframes rotate {
-  100% {
-    transform: rotate(360deg);
-  }
-}
-
-@keyframes dash {
-  0% {
-    stroke-dasharray: 1, 150;
-    stroke-dashoffset: 0;
-  }
-
-  50% {
-    stroke-dasharray: 90, 150;
-    stroke-dashoffset: -35;
-  }
-
-  100% {
-    stroke-dasharray: 90, 150;
-    stroke-dashoffset: -124;
-  }
-}
