+++ src/components/SelectCustom/Chain/ChainUX.module.css
@@ -13,7 +13,7 @@
 .name {
   font-family: "Bai Jamjuree", sans-serif;
   font-weight: 400;
-  font-size: 15px;
+  font-size: 14px;
   margin: 0;
   color: white;
   text-align: left;
