+++ src/components/Swap/Swap.module.css
@@ -2,7 +2,7 @@
   display: flex;
   flex-direction: column;
   align-items: center;
-  margin: 150px auto;
+  margin: 150px auto 76px;
 }
 
 .swap {
@@ -65,7 +65,6 @@
 }
 
 .selectWrapper:hover {
-  background-color: #e2fb5f;
   background: rgba(67, 151, 178, 0.1);
 }
 
@@ -80,7 +79,7 @@
 }
 
 .circle:hover {
-  background-color: #EDFF8E;
+  background-color: #bbd24d;
 }
 
 .line {
