+++ src/components/WelcomePage/BasicWindow/BasicWindow.tsx
@@ -11,6 +11,8 @@ const BasicWindow = () => {
         <div className={styles.WelcomeBlockSecond}>
           create! <br /> manage! <br /> swap!
         </div>
+      </div>
+      <div className={styles.mainPageWelcomeBlockinerInner}>
         <div className={styles.WelcomeBlockThird}>
           All-in-One <br /> token hub
         </div>
